requirements.txt
pyelftools==0.32
stickytape==0.2.1
